--- a/api/proto/teleport/userloginstate/v1/userloginstate.proto
+++ b/api/proto/teleport/userloginstate/v1/userloginstate.proto
@@ -37,4 +37,7 @@ message Spec {
 
   // traits are the traits attached to the user.
   repeated teleport.trait.v1.Trait traits = 2;
+
+  // user_type is the type of user this state represents.
+  string user_type = 3;
 }