--- a/lib/auth/auth.go
+++ b/lib/auth/auth.go
@@ -63,6 +63,7 @@ import (
 	"github.com/gravitational/teleport/api/internalutils/stream"
 	"github.com/gravitational/teleport/api/types"
 	apievents "github.com/gravitational/teleport/api/types/events"
+	apiuserloginstate "github.com/gravitational/teleport/api/types/userloginstate"
 	"github.com/gravitational/teleport/api/types/wrappers"
 	apiutils "github.com/gravitational/teleport/api/utils"
 	"github.com/gravitational/teleport/api/utils/keys"
@@ -410,7 +411,7 @@ func NewServer(cfg *InitConfig, opts ...ServerOption) (*Server, error) {
 	// Add in a login hook for generating state during user login.
 	ulsGenerator, err := userloginstate.NewGenerator(userloginstate.GeneratorConfig{
 		AccessLists: services,
-		Roles:       services,
+		Access:      services,
 		Clock:       cfg.Clock,
 	})
 	if err != nil {
@@ -2342,20 +2343,9 @@ func generateCert(a *Server, req certRequest, caType types.CertAuthType) (*proto
 
 	// Add in the user login state if it exists.
 	uls, err := a.GetUserLoginState(ctx, req.user.GetName())
-	if err == nil {
-		// Copy roles to the roles from the request.
-		roles = append(roles, uls.GetRoles()...)
-
-		// Copy traits to the traits from the request.
-		if len(uls.GetTraits()) > 0 {
-			if traits == nil {
-				traits = map[string][]string{}
-			}
-			for k, v := range uls.GetTraits() {
-				traits[k] = append(traits[k], v...)
-			}
-		}
-	} else if !trace.IsNotFound(err) {
+	if uls != nil {
+		traits, roles = addUserLoginStateToTraitsAndRoles(uls, traits, roles)
+	} else if err != nil && !trace.IsNotFound(err) {
 		return nil, trace.Wrap(err)
 	}
 
@@ -2559,6 +2549,35 @@ func generateCert(a *Server, req certRequest, caType types.CertAuthType) (*proto
 	a.submitCertificateIssuedEvent(&req)
 
 	return certs, nil
+}
+
+// addUserLoginStateToRolesAndTraits will modify roles and traits by adding in the user login state to them, making
+// sure to deduplicate as well.
+func addUserLoginStateToTraitsAndRoles(uls *apiuserloginstate.UserLoginState, traits map[string][]string, roles []string) (map[string][]string, []string) {
+	if uls == nil {
+		return traits, roles
+	}
+
+	// Copy traits to the traits from the request.
+	if len(uls.GetTraits()) > 0 {
+		if traits == nil {
+			traits = map[string][]string{}
+		}
+		for k, v := range uls.GetTraits() {
+			traits[k] = append(traits[k], v...)
+		}
+
+		for k, v := range traits {
+			traits[k] = apiutils.Deduplicate(v)
+		}
+	}
+
+	// Copy roles to the roles from the request and deduplicate.
+	if len(uls.GetRoles()) > 0 {
+		roles = apiutils.Deduplicate(append(roles, uls.GetRoles()...))
+	}
+
+	return traits, roles
 }
 
 type verifyLocksForUserCertsReq struct {