--- a/lib/auth/auth.go
+++ b/lib/auth/auth.go
@@ -410,6 +410,7 @@ func NewServer(cfg *InitConfig, opts ...ServerOption) (*Server, error) {
 
 	// Add in a login hook for generating state during user login.
 	ulsGenerator, err := userloginstate.NewGenerator(userloginstate.GeneratorConfig{
+		Log:         log,
 		AccessLists: services,
 		Access:      services,
 		Clock:       cfg.Clock,