… api/gen/proto/go/teleport/userloginstate/v1/userloginstate.pb.go
… api/types/userloginstate/convert/v1/user_login_state_test.go
… lib/auth/auth_test.go
… lib/auth/auth_with_roles.go
… lib/auth/bot.go
… lib/auth/helpers.go
… lib/auth/init.go
… lib/auth/methods.go
… lib/auth/sessions.go
… lib/auth/userloginstate/generator.go
… lib/auth/userloginstate/generator_test.go
… lib/auth/userloginstate/service_test.go
… lib/services/access_checker.go
… lib/services/parser.go
… lib/services/role_test.go
… lib/web/app/handler.go