On login, the user login state will be generated, using access lists to register additional roles and traits that will be inserted into the user's certificate.

Tests have been added to exercise this as well.

[Relevant section in the RFD](https://github.com/gravitational/teleport.e/blob/master/rfd/0006e-access-lists.md#userloginstate-object).

Small low key benchmark:

```
With login state

AuthenticateUser begin: 2023-07-20 13:35:22.553049 -0400 EDT m=+23.786587251
AuthenticateUser end: 2023-07-20 13:35:22.66828 -0400 EDT m=+23.901817918
AuthenticateUser delta: 115.230667ms
GenerateCert begin: 2023-07-20 13:35:22.668675 -0400 EDT m=+23.902212501
GenerateCert end: 2023-07-20 13:35:22.673191 -0400 EDT m=+23.906728626
GenerateCert delta: 4.516125ms

Without login state

AuthenticateUser begin: 2023-07-20 13:40:13.515064 -0400 EDT m=+12.754441292
AuthenticateUser end: 2023-07-20 13:40:13.636613 -0400 EDT m=+12.875989501
AuthenticateUser delta: 121.548209ms
GenerateCert begin: 2023-07-20 13:40:13.63706 -0400 EDT m=+12.876436501
GenerateCert end: 2023-07-20 13:40:13.641253 -0400 EDT m=+12.880629459
GenerateCert delta: 4.192958ms

With login state + 2 access lists (not a member)

AuthenticateUser begin: 2023-07-20 13:50:17.48219 -0400 EDT m=+126.730266376
AuthenticateUser end: 2023-07-20 13:50:17.595715 -0400 EDT m=+126.843790251
AuthenticateUser delta: 113.523875ms
GenerateCert begin: 2023-07-20 13:50:17.596153 -0400 EDT m=+126.844228043
GenerateCert end: 2023-07-20 13:50:17.60055 -0400 EDT m=+126.848625126
GenerateCert delta: 4.397083ms

With login state + 2 access lists (member)

AuthenticateUser begin: 2023-07-20 14:05:50.527176 -0400 EDT m=+141.383355001
AuthenticateUser end: 2023-07-20 14:05:50.633066 -0400 EDT m=+141.489245043
AuthenticateUser delta: 105.890042ms
GenerateCert begin: 2023-07-20 14:05:50.633598 -0400 EDT m=+141.489777001
GenerateCert end: 2023-07-20 14:05:50.637755 -0400 EDT m=+141.493933918
GenerateCert delta: 4.156917ms
```