@mdwn 2023-07-21
> As I understood the original discussion about this feature, user login state was intended to be a complete representation of the "post-mapping" user-state that we could use as a drop-in replacement for the user resource as-appropriate. As currently implemented, the various cert-generation methods are still loading the user resource as the primary source of truth, and then using the login state to extend that. So currently we end up with a pattern like:
> 
> ```go
> 	user, _ := a.GetUser(req.Username, false)
> 	// ...
> 	accessInfo := services.AccessInfoFromUser(user)
> 	// ...
> 	checker, _ := services.NewAccessChecker(accessInfo, ...)
> 	// ...
> 	traits, roles = addUserLoginStateToTraitsAndRoles(uls, req.traits, req.checker.RoleNames())
> 	// ...etc
> ```
> 
> Which means that we're still mixing pre-login and post-login state (e.g. the access checker is initialized with your pre-mapping state).
> 
> I don't think this is how we should be going about it. I think we'd be better off migrating to the login state as the new source of truth, and tweaking things like `AccessChecker` to work with an interface that abstracts over the different between user state and user. Then we could do something like this (oversimplified):
> 
> ```go
> func getUserOrLoginState(name string) UserStateInterface {
>     if uls, _ := getUserLoginState(name); uls != nil {
>         return uls
>     }
>     return getUser(name)
> }
> 
> // ...
> 
> accessInfo := services.AccessInfoFromUserStateInterface(getUserOrLoginState())
> 
> // remaining logic is totally identical to what we had before
> ```
> 
> Basically, I feel like it's going to cause issues if we sometimes work from the pre-mapping state and sometimes from the post-mapping state. Much better to work exclusively with the post-mapping state, and fall back to the pre-mapping state of the post mapping state does not exist (e.g. due to login having been handled by an older auth server).
> 
> WDYT? Does this sound reasonable/doable?

I think this sounds pretty reasonable. I don't think it'll be significantly more work than what I've already got here, it'll just involve starting with the "user" as the base user login state and then going from there. I'll update when I've made these changes.